+++ src/biosphere_pmodel.mod.f90
@@ -135,7 +135,7 @@ contains
                   tile_fluxes(:), &
                   myinterface%pco2, &
                   myinterface%climate(doy), &
-                  myinterface%grid, &
+                  ! myinterface%grid, & (unused)
                   init_daily, &
                   myinterface%params_siml%in_ppfd, &
                   myinterface%tc_home &